Create starter dodo.py
This is the main module for running doit. All contents here tried to
reproduce what the old Makefile was doing, so we migrate from GNU Make
to doit.

Some functions still need tweaking and may be changed later. This is a
prototype. However, some functions are pretty good and may stay.

The location of ACS files was put in an 'external' directory. This is
beacuse they come from CPE now. May change to 'raw' later, when
automated.