Create Department.__new__()
Some departments (all of them?) will need custom logic. This commit
changes the Department constructor so that it returns a specific
subclass, based on the department name.

If no specific subclass was found, we stil use the default Department.
(may change this behavior later for a custom subclass with default
behavior)... Naah, we want to use OOP to the max here.