Add path properties for Department (#7)
In lots of places in the code, we were using dept.dir / 'dirA' / 'dirB' / 'filname.ext'. While this approach is clearly visible in what it is achieving, it lacks consistency. Also, it makes it difficult to change paths later. As, to change the path of one single file, one would have to touch lots of different places of the code.

To solve this, I created custom properties for the Department class. The path properties in the Department class itself are gonna be those that are shared between every department. For single departments (subclasses), there may also be some other path properties, and they may override the original properties too.