Add IO methods for Department
While loading/saving files directly is transparent, we want a more
contained method, so that it can be modified in subclasses and is
consistent and less prone to errors. Therefore, we abstract the input
and output later into different methods named load_{file} and
save_{file}.

The same scheme can also be used in the census and I believe is a common
programming thing.