Create task_create_dept_list and related (#9)
This commit defines a 'create_dept_list' task and all its associated stuff.

For example, DepartmentColl class was create to handle things that should be done on a whole, that is, for all departments at once. We still give preference to Department for things that can be split into each one though.