IO zipshp
Lots of trouble here... We want to use the root of a zipfile to store
the (shapefile) files. However if we try loading the root as
gpd.read_file('/', vfs='zip://shapes.zip'), fiona will raise an
Exception. I believe this is a bug, but also may be a usage
misconception.

To fix this, we can always store the files in a different VFS dir, e.g.
gpd.read_file('/shapefile', vfs='zip://shapes.zip'). While it works,
this won't work well for external data retrieved from websites, as the
shapefiles are usually (up until now) found in the root dir.

So, the solution I found (looks undocumented) was to use the zip URI
directly. I don't know if this was intended behavior, but it works, and
we are gonna roll with it. For example,
gpd.read_file('zip://shapes.zip') just work!