Use zipped shapefiles instead of dirs (#10)
Lots of things happening here...

Most importantly, I'm giving preference to zipped shapefiles and this
is gonna be the usual format for this project. This decision came with
a lot of trouble in using other spatial storages... The reasons for it are:

- Some of the data (e.g. ACS) comes in this format. That means no need
  to convert or unzip.
- It is just one step away from the standard shapefile, so, shall provide
  no trouble to the folks at CPE.
- It is contained in only one file. This is essential for keeping file
  dependencies with doit.

There was also some linting done and the code looks way clean
now.