Put preparation tasks into another module
Preparing the outputs in the dodo.py module lead us to an excessive use
of the doit.create_after decorator. With this, the generated targets
were not being registered (because they were delayed) and the order of
the tasks get messy.

To solve this, I just split the prepartion tasks and the main taks into
two differet modules. Before anything in the main module must be done,
you must first run the preparation module.