Create ACS helper module
This commit creates the ACS class that will be used to retrieve data
from the ACS API. Mainly, you will want to import get_acs() from
__init__.py and then call the method ACS.data() to retrieve your data.

Code is linted. We will want to change the behavior of census.py later
and isolate it into the TIGER related thing or join it with the acs.py
module (code and/or functionality (see walkerke/tidycensus)).

Some census configuration have also been added, these must be merged
into the TIGER utilities later. And the get_cnensus or get_tiger is
coming! Just need some more walking on the project to check how it will
be designed.