Fix GeoJSON overwrite problem
Fiona (and, consequently, geopandas) doesn't like to write on top of an
existing GeoJSON file. This would cause problems when and file in the
doit pipeline should be updated.

To solve the issue, I made the IO of shapefiles and GeoJSON functions of
the util.io module. Now, we can abstract further from geopandas and the
file problem could be solved easily by always removing a file before
save (we may want to be more careful, though, and write to a temporary
location before erasing the contents of the old file).