Choose shapefile for Department 35-00016
The data provided for Department 35-00016 (Orlando, FL) consisted of
three shapefiles: Police Divisions, Police Sectors and Police Districts.
It is a crucial step to choose only one to continue working with,
therefore, I choose Police Districts.

This commit implements the load of the police districts only, in a
custom load_external_shapefile method (I think the default was already
loading districts, but it is better to make the choice explicit).