Retrieve new ACS variables
As we go on with the analysis, the need for new ACS variables appear. I
tried using everything that seems to be used by CPE right now. Also,
'Total Population' has been renamed to 'TOTAL_POPULATION'.

The naming of the variables is composed of a subject and the variable
itself. The subject helps limit scope and makes variables easier to
find. For example, the variable representing the total amount of
people of hispanic origin is named 'RACE_HISPANIC'.

Groups of variables added:

- Sex (male/female)
- Race (white/black/etc)
- Poverty (below poverty level)
- Employment (employed)
- ~~Education (less than HS/HS diploma/etc)~~

Education was not added because the table containing its values had
*lots* of variables (B15003). There were other tables with summarized
information, but, they were not available at BG level.