Preprocess department 49-00033 arrests (#15)
Implement an ad-hoc preprocessing for the Department 49-00033 arrests
table (actually I'm not sure if all arrests or just incidents).

The need for this preprocessing became evident in the previously done
EDA, and, I believe it will be useful for CPE.

Later on, we will need to abstract this into more departments,
otherwise, the code may get overwhelmingly big and fragmented. Let's see
what to do!