Set up input scheme (#17)
Stop the retrieval and processing of files directly from Kaggle. The
files are now present in the repo itself. Also change directories names
and set up input scheme to data/input/department/*/<format>, where
format can be either tabular or spatial.

The Kaggle change was done to simplify the process for CPE. They don't
wanna see how code was retrieved and processed from Kaggle, they just
need to see where they should put files and the outputs that they might
expect. This is much easier, and, moreover, retrieving the data from
Kaggle required a Kaggle account, which in itself was gonna be a hassle
to the usage of the package.

Now, for the file changes, we changed the whole scheme of inputs to
reflect a was that feels better. By clearly defining the utility as a
file processor, we see that the usage of an `output` and `input` is
superior to that approach that is only object-centered. The tasks become
clearly defined.

Also, please note that this is in an alpha stage, and, as more
modifications are being made (specially when we implement the output
functions), some of what was made here will change. I still believe this
was a robust PR, but I also think there will be the need for cleaning
everything later on (specially `prepare.py` and
`cpe_help/department.py`)