Remove the ensure_path function
Since the directories refactor, we are not using it anymore.

It was good using it for ensuring that a certain directory exists...
However, by always doing this, it's like we admit there is a lot of
uncertainty in the software and we don't know where directories are
gonna be present or not.

Also, there was the problems that some of the save functions were from
outer packages and *did not* ensured the path to a certain file existed.
This caused lots of problems and leads to hacking those packages or
creating utilities just so that this unsatisfying condition must be met.

A clearer approach results in more robust designing and a clearer tree.