Set up CI with CircleCI
The CI creates a conda environment, installs the project in edit mode
(pip install -e) and then attempts an import for the root package.

The install may be not necessary... We may also want to add tests later.
The import is just to catch unexpected imports errors, as the root
packages surfaces a big part of the code.