Split preparation steps (#30)
Split doit preparation steps into two workflows:

- Manual preparation, where you manually create department directories and put the files
- Automatic preparation, where the department directories are created directly from the Kaggle input

Also allow for the combination of the two workflows together.

To clean the repo, Kaggle contents are now saved zipped, and unzipped when they are actually needed.

Also, data/input directory is now a part of the repo, as it will be directly used by CPE to put files in.

The need of this PR was because I got started to think about the CPE workflow. How should they work with files and such. With the current method now, it is possibile for them to do things much easier. The last one was complicated and hard to understand and mixed Kaggle and manual process (not even sure if it would work correctly).